[UPDATE] mongodb 재구조
