+++ app/core/data/mongo/token/__init__.py
@@ -1,0 +1,10 @@
+class MessengerToken(dict):
+    TOKEN = '_id'
+
+    def __init__(
+            self,
+            token: str
+    ):
+        super().__init__(
+            _id=token
+        )
