+++ app/core/data/mongo/token/query.py
@@ -1,5 +1,5 @@
 from app.core.data.mongo import CollectionNames, get_collection
-from app.core.data.mongo.collections import MessengerToken
+from app.core.data.mongo.token import MessengerToken
 
 
 def get_page_access_token():
