+++ app/core/github/logic/event/non_operate.py
@@ -1,7 +1,7 @@
 from uuid import UUID
 from fastapi import Request
 
-from app.core.data.mongo.collections import Webhook, WebhookUser, Repository
+from app.core.data.mongo.webhook import Webhook, User, Repository
 from app.core.data.mongo.token.query import get_page_access_token
 from app.core.data.mongo.webhook.query import get_webhook_by_id
 from app.core.data.request.messenger import send_message
@@ -15,6 +15,6 @@ async def handle_none_provide_event(bot_id: UUID, event_type: str, request: dict
 
     send_message(
         get_page_access_token(),
-        webhook[Webhook.USER][WebhookUser.SENDER_ID],
+        webhook[Webhook.USER][User.SENDER_ID],
         message
     )
